+++ package.json
@@ -111,6 +111,7 @@
         "react-test-renderer": "^16.8.6",
         "sass": "^1.26.10",
         "sass-loader": "^9.0.2",
+        "speed-measure-webpack-plugin": "^1.5.0",
         "style-loader": "^1.2.1",
         "stylelint": "^9.10.1",
         "stylelint-config-concentric": "^2.0.2",
